Fix tests, detect bogus arguments

+++ spec/hace_spec.cr
@@ -7,7 +7,7 @@ def with_scenario(name, keep = [] of String, &)
   # ::logs = IO::Memory.new
   # Log.setup(:debug, Log::IOBackend.new(io: logs, formatter: Log::ShortFormat))
   Dir.cd("spec/testcases/#{name}") do
-    File.delete?(".croupier")
+    File.delete?(".croupier") unless keep.includes? ".croupier"
     Dir.glob("*").each do |f|
       File.delete?(f) unless f == "Hacefile.yml" || keep.includes?(f)
     end
@@ -109,23 +109,6 @@ describe Hace do
       end
     end
 
-    it "should do nothing when running a second time" do
-      with_scenario("run_all") do
-        File.open("bar", "w") do |io|
-          io << "1111"
-        end
-        HaceFile.run # This should put "1111" in "foo"
-        TaskManager.tasks.values.select(&.stale?).empty?.should be_true
-        File.read("foo").should eq "1111"
-        # This should NOT put "2222" in "foo" because the tasks have ran
-        File.open("bar", "w") do |io|
-          io << "2222"
-        end
-        HaceFile.run
-        File.read("foo").should eq "1111"
-      end
-    end
-
     it "should run normally when running a second time with run_all=true" do
       with_scenario("run_all") do
         File.open("bar", "w") do |io|
@@ -138,6 +121,7 @@ describe Hace do
         File.open("bar", "w") do |io|
           io << "2222"
         end
+        TaskManager.cleanup
         HaceFile.run(run_all: true)
         File.read("foo").should eq "2222"
       end
